+++ packages/@leight/job/src/schema/JobQuerySchema.ts
@@ -1,13 +1,10 @@
-import {
-    JobOrderByWithRelationInputSchema,
-    JobWhereInputSchema
-}                          from "@leight/prisma";
+import {Schema}            from "@leight/prisma";
 import {createQuerySchema} from "@leight/query";
 import {z}                 from "zod";
 
 export const JobQuerySchema = createQuerySchema({
-    filter: JobWhereInputSchema,
-    sort:   JobOrderByWithRelationInputSchema,
+    filter: Schema.JobWhereInputSchema,
+    sort:   Schema.JobOrderByWithRelationInputSchema,
 });
 
 export type IJobQuery = z.infer<typeof JobQuerySchema>;