+++ packages/@leight/user-server/src/service/RegistrationService.ts
@@ -22,10 +22,11 @@ export class RegistrationService implements IRegistrationService {
     ) {
     }
 
-    public async handle<T extends IToken>({
-                                              token,
-                                              isNewUser,
-                                          }: IRegistrationService.HandleProps<T>): Promise<void> {
+    public async handle<T extends IToken>(
+        {
+            token,
+            isNewUser,
+        }: IRegistrationService.HandleProps<T>): Promise<void> {
         if (!isNewUser || !token.sub) {
             return;
         }
@@ -39,24 +40,26 @@ export class RegistrationService implements IRegistrationService {
     /**
      * Define defaults for the new root user (tokens and co.).
      */
-    private async registerRootUser<T extends IToken>({
-                                                         token: {sub},
-                                                     }: IRegistrationService.HandleProps<T>): Promise<void> {
+    private async registerRootUser<T extends IToken>(
+        {
+            token: {sub},
+        }: IRegistrationService.HandleProps<T>): Promise<void> {
         const user = await this.prisma.user.findUniqueOrThrow({
             where: {id: sub},
         });
-        user.id = "user.id";
+        user.id    = "user.id";
     }
 
     /**
      * Define defaults for the new common user (tokens and co.).
      */
-    private async registerCommonUser<T extends IToken>({
-                                                           token: {sub},
-                                                       }: IRegistrationService.HandleProps<T>): Promise<void> {
+    private async registerCommonUser<T extends IToken>(
+        {
+            token: {sub},
+        }: IRegistrationService.HandleProps<T>): Promise<void> {
         const user = await this.prisma.user.findUniqueOrThrow({
             where: {id: sub},
         });
-        user.id = "user.id";
+        user.id    = "user.id";
     }
 }