+++ packages/@leight/i18n-client/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @leight/i18n-client
 
+## 0.5.25
+
+### Patch Changes
+
+- Just a record here
+- Updated dependencies
+    - @leight/context-client@0.5.30
+    - @leight/utils@0.5.26
+    - @leight/i18n@0.5.24
+
+## 0.5.24
+
+### Patch Changes
+
+- Version bump II
+- Updated dependencies
+    - @leight/context-client@0.5.29
+    - @leight/utils@0.5.25
+    - @leight/i18n@0.5.23
+
+## 0.5.23
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/context-client@0.5.28
+    - @leight/i18n@0.5.22
+    - @leight/utils@0.5.24
+
 ## 0.5.22
 
 ### Patch Changes