+++ packages/@leight/i18n-server/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @leight/i18n-server
 
+## 0.5.24
+
+### Patch Changes
+
+- Just a record here
+- Updated dependencies
+    - @leight/i18n@0.5.24
+
+## 0.5.23
+
+### Patch Changes
+
+- Version bump II
+- Updated dependencies
+    - @leight/i18n@0.5.23
+
+## 0.5.22
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/i18n@0.5.22
+
 ## 0.5.21
 
 ### Patch Changes