+++ packages/@leight/import-server/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @leight/import-server
 
+## 0.5.23
+
+### Patch Changes
+
+- Just a record here
+- Updated dependencies
+    - @leight/container@0.5.23
+    - @leight/import@0.5.23
+
+## 0.5.22
+
+### Patch Changes
+
+- Version bump II
+- Updated dependencies
+    - @leight/container@0.5.22
+    - @leight/import@0.5.22
+
+## 0.5.21
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/container@0.5.21
+    - @leight/import@0.5.21
+
 ## 0.5.20
 
 ### Patch Changes