+++ packages/@leight/next.js-server/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @leight/next.js-server
 
+## 0.5.24
+
+### Patch Changes
+
+- Just a record here
+- Updated dependencies
+    - @leight/user-server@0.5.23
+    - @leight/container@0.5.23
+    - @leight/winston@0.5.23
+    - @leight/utils@0.5.26
+    - @leight/user@0.5.24
+
+## 0.5.23
+
+### Patch Changes
+
+- Version bump II
+- Updated dependencies
+    - @leight/user-server@0.5.22
+    - @leight/container@0.5.22
+    - @leight/winston@0.5.22
+    - @leight/utils@0.5.25
+    - @leight/user@0.5.23
+
+## 0.5.22
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/container@0.5.21
+    - @leight/user@0.5.22
+    - @leight/user-server@0.5.21
+    - @leight/utils@0.5.24
+    - @leight/winston@0.5.21
+
 ## 0.5.21
 
 ### Patch Changes