+++ packages/@leight/sql/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @leight/sql
 
+## 0.5.23
+
+### Patch Changes
+
+- Just a record here
+- Updated dependencies
+    - @leight/prisma@0.5.26
+
+## 0.5.22
+
+### Patch Changes
+
+- Version bump II
+- Updated dependencies
+    - @leight/prisma@0.5.25
+
+## 0.5.21
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.24
+
 ## 0.5.20
 
 ### Patch Changes