+++ packages/@leight/trpc-server/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @leight/trpc-server
 
+## 0.5.24
+
+### Patch Changes
+
+- Just a record here
+- Updated dependencies
+    - @leight/user-server@0.5.23
+    - @leight/container@0.5.23
+    - @leight/trpc@0.5.23
+    - @leight/user@0.5.24
+
+## 0.5.23
+
+### Patch Changes
+
+- Version bump II
+- Updated dependencies
+    - @leight/user-server@0.5.22
+    - @leight/container@0.5.22
+    - @leight/trpc@0.5.22
+    - @leight/user@0.5.23
+
+## 0.5.22
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/container@0.5.21
+    - @leight/trpc@0.5.21
+    - @leight/user@0.5.22
+    - @leight/user-server@0.5.21
+
 ## 0.5.21
 
 ### Patch Changes