+++ packages/@leight/utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @leight/utils
 
+## 0.5.26
+
+### Patch Changes
+
+- Just a record here
+
+## 0.5.25
+
+### Patch Changes
+
+- Version bump II
+
+## 0.5.24
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.23
 
 ### Patch Changes