+++ packages/@leight/xlsx-import/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @leight/xlsx-import
 
+## 0.5.24
+
+### Patch Changes
+
+- Just a record here
+- Updated dependencies
+    - @leight/import@0.5.23
+    - @leight/trpc@0.5.23
+    - @leight/job@0.5.25
+
+## 0.5.23
+
+### Patch Changes
+
+- Version bump II
+- Updated dependencies
+    - @leight/import@0.5.22
+    - @leight/trpc@0.5.22
+    - @leight/job@0.5.24
+
+## 0.5.22
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/import@0.5.21
+    - @leight/job@0.5.23
+    - @leight/trpc@0.5.21
+
 ## 0.5.21
 
 ### Patch Changes